Refactor: note class to new folder
